requirements.txt
bcrypt==3.1.1
cffi==1.9.1
click==6.6
decorator==4.0.10
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-Markdown==0.3
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8
Markdown==2.6.7
MarkupSafe==0.23
pbr==1.10.0
pycparser==2.17
six==1.10.0
SQLAlchemy==1.1.3
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
Werkzeug==0.11.11
WTForms==2.1
